Correcting FSLogixAppsSetup.exe directory location
Download file from https://aka.ms/fslogix_download no longer has a parent folder for the x64 & Win32 directories, install fails due to not finding the correct path to Setup.exe. Adding check to test if there is a subfolder or not, and if so grab the correct parent directory.