Update uninstall logic, install Teams Meeting add-in
Update the package uninstall logic to find the product code based on the package name. This will keep changes in the product code from breaking the uninstall portion of the script.

This update will also check for the Teams Meeting add-in and uninstall it if it is present, then reinstall the latest version. The meeting add-in is no longer bundled by default in the Teams installer, and must instead be installed manually from the MSIX package.